bug 1302459: Move MDN_LANGUAGES to languages.py
Move MDN_LANGUAGES and LOCALES to languages.py, but still load from the
product details JSON. This will eventually be a generated, but
checked-in file.
